refactor layout management by removing deprecated layout constants and restructuring IPC calls for workspace and layout state
