--- view/desktop/src/constants/layout.ts
@@ -1,19 +1,0 @@
-// Layout constants
-
-// Sidebar related constants
-export const ACTIVITY_BAR_WIDTH = 41;
-export const DEFAULT_SIDEBAR_WIDTH = 270;
-export const MIN_SIDEBAR_WIDTH = 100;
-export const MAX_SIDEBAR_WIDTH = 400;
-export const SIDEBAR_COLLAPSE_THRESHOLD = 105;
-
-// Bottom pane related constants
-export const MIN_BOTTOM_PANE_HEIGHT = 100;
-export const MIN_BOTTOM_PANE_DRAGGABLE_HEIGHT = 100;
-export const DEFAULT_BOTTOM_PANE_HEIGHT = 250;
-export const BOTTOM_PANE_COLLAPSE_THRESHOLD = 105;
-
-// Sidebar position constants
-export const SIDEBAR_POSITION_LEFT = "left";
-export const SIDEBAR_POSITION_RIGHT = "right";
-export const SIDEBAR_POSITION_NONE = "none";
