+++ view/desktop/src/tauriApi/index.ts
@@ -1,0 +1,3 @@
+export * from "./describeLayoutPartsState";
+export * from "./openWorkspace";
+export * from "./setLayoutPartsState";
