+++ view/desktop/src/tauriApi/openWorkspace.ts
@@ -1,0 +1,8 @@
+import { invokeTauriIpc } from "@/lib/backend/tauri";
+import { OpenWorkspaceInput, OpenWorkspaceOutput } from "@repo/moss-workspace";
+
+export const openWorkspace = async (name: string) => {
+  return await invokeTauriIpc<OpenWorkspaceInput, OpenWorkspaceOutput>("open_workspace", {
+    input: { name },
+  });
+};
